Add support for the compiler -S flag (emit assembly to file only)

+++ source/driver/driver.cpp
@@ -17,7 +17,7 @@ void runDriver(std::vector<std::string>& arguments) {
         const auto preprocessed = runPreprocessor(args);
         const auto [proceed, assembly] = runCompiler(preprocessed, args);
 
-        if (proceed) {
+        if (proceed && !args.special) {
             runAssembler(assembly, args);
         }
 
@@ -34,19 +34,20 @@ void cleanUp(const DriverArgs& args, bool failed) {
     const auto info = utils::getFileInfo(args.path);
     auto source = std::filesystem::path{args.path};
     if (!std::filesystem::exists(source)) return;
-    
+
     auto directory = std::filesystem::directory_iterator{info.parent};
 
-    if (failed) {
-        for (auto& file : directory) {
-            const auto& path = file.path();
+    for (auto& file : directory) {
+        const auto& path = file.path();
+        if (args.special && path.string().ends_with(".s")) {
+            continue;
+        }
+        
+        if (failed) {
             if (path != source) {
                 std::filesystem::remove(path);
             }
-        }
-    } else {
-        for (auto& file : directory) {
-            const auto& path = file.path();
+        } else {
             if (path.has_extension() && path != source) {
                 std::filesystem::remove(path);
             }
